feat: Integrate slicing progress directly into plate selector with persistent estimates
- Move streaming slice logic from StreamingSliceTracker into PlateSelector component
- Add real-time progress overlays directly on plate thumbnail cards
- Implement local cache for plate estimates to prevent loss during multi-plate slicing
- Create subtle progress fill animation (5%/3% opacity) with completion fade-out
- Remove standalone StreamingSliceTracker component as functionality is now integrated
- Fix issue where plate estimates reverted to "After slice" when other plates completed
- Add proper cleanup of EventSource connections and progress state management
